Slight adjustments to the script...
Turns put that DNF was confusing the app with ANOTHER app that has a similar name, so if one of you would decide to update your system, DNF will remove the app and "update" it with GIS software. I also fixed the guideline command at the end of the script. (sudo dnf localinstall is not a valid command)